Added Task 3 Files
Added Task 3 files
Task 3: Configure the inverted pendulum simulator to create two inverted
pendulums on the same track (See the README file in the simulator on how
to configure). Then, design the control algorithms for both of the
pendulums so that the first pendulum should move from the initial
position to a target position; and the second pendulum should follow the
first pendulum (with a distance between them to avoid collision).

The control algorithm of the first pendulum is the same as the one
developed in Task 2. The control algorithm of the second pendulum should
not use the target position of the first pendulum, because it just
follows the first pendulum and does not know its planned trajectory. It
should only use the sensing data of the first pendulum.